+++ packages/guardian/src/index.ts
@@ -1,3 +1,7 @@
+import pkg from "../package.json"
+
+export const VERSION = pkg.version
+
 export * from "./domain"
 export * from "./application"
 export * from "./infrastructure"